gemini-legion-backend/requirements.txt
fastapi
uvicorn[standard]
google-adk
python-dotenv
google-generativeai
httpx
